get orbits to work

+++ computation/Star.py
@@ -1,4 +1,4 @@
 import taichi as ti
 import taichi.math as tm
 
-Star = ti.types.struct(pos=tm.vec3, radius=ti.f32, m=ti.u64, v=tm.vec3, f=tm.vec3)
+Star = ti.types.struct(pos=tm.vec3, radius=ti.u32, m=ti.u64, v=tm.vec3, f=tm.vec3)
